降低了非平行线段减法误差带来的影响，降低了第一步分组错误带来的部分影响（不可能车道），暂时就这样了啊。

+++ include/KMeanClassify.h
@@ -77,7 +77,7 @@ bool KMeanClassify(const vector<ElementType> &element_vec, vector<vector<Element
 				continue;
 			}
 			ElementType lineSegment_curr = element_vec[i];
-			int min_distance_index = 0;
+			size_t min_distance_index = 0;
 			double min_value = INT_MAX;
 			double value;
 			for (size_t j = 0; j < types; ++j)
